Move enum definitions to a separate file
